requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
alembic==1.13.1
    # via -r requirements.in
amqp==5.2.0
    # via kombu
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-asgi-testclient==1.4.11
    # via -r requirements.in
asynch==0.2.3
    # via -r requirements.in
asyncmy==0.2.9
    # via -r requirements.in
asyncpg==0.29.0
    # via -r requirements.in
bcrypt==4.1.2
    # via passlib
billiard==4.2.0
    # via celery
build==1.1.1
    # via pip-tools
celery==5.3.6
    # via
    #   -r requirements.in
    #   flower
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
ciso8601==2.3.1
    # via asynch
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
deprecation==2.1.0
    # via testcontainers
distlib==0.3.8
    # via virtualenv
dnspython==2.6.1
    # via email-validator
docker==7.0.0
    # via
    #   testcontainers
    #   testcontainers-core
ecdsa==0.18.0
    # via python-jose
email-validator==2.1.1
    # via pydantic
factory-boy==3.3.0
    # via -r requirements.in
faker==24.1.0
    # via factory-boy
fastapi==0.109.0
    # via -r requirements.in
filelock==3.13.1
    # via virtualenv
flower==2.0.1
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via -r requirements.in
humanize==4.9.0
    # via flower
identify==2.5.35
    # via pre-commit
idna==3.6
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
kombu==5.3.5
    # via celery
leb128==1.0.6
    # via asynch
lz4==4.3.3
    # via asynch
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
multidict==6.0.5
    # via async-asgi-testclient
nodeenv==1.8.0
    # via pre-commit
packaging==23.2
    # via
    #   build
    #   deprecation
    #   docker
    #   pytest
passlib[bcrypt]==1.7.4
    # via
    #   -r requirements.in
    #   passlib
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.2.0
    # via virtualenv
pluggy==1.4.0
    # via pytest
pre-commit==3.6.0
    # via -r requirements.in
prometheus-client==0.20.0
    # via flower
prompt-toolkit==3.0.43
    # via click-repl
psycopg==3.1.16
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via testcontainers-postgres
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pydantic[email]==2.6.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic
    #   pydantic-settings
pydantic-core==2.16.1
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.23.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   faker
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
pytz==2024.1
    # via
    #   asynch
    #   flower
pyyaml==6.0.1
    # via
    #   pre-commit
    #   uvicorn
redis==5.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   async-asgi-testclient
    #   docker
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.25
    # via
    #   -r requirements.in
    #   alembic
    #   testcontainers-postgres
starlette==0.35.1
    # via fastapi
testcontainers==3.7.1
    # via -r requirements.in
testcontainers-core==0.0.1rc1
    # via testcontainers-postgres
testcontainers-postgres==0.0.1rc1
    # via -r requirements.in
tornado==6.4
    # via flower
typing-extensions==4.10.0
    # via
    #   alembic
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via celery
tzlocal==5.2
    # via asynch
urllib3==2.2.1
    # via
    #   docker
    #   requests
uvicorn[standard]==0.27.0
    # via
    #   -r requirements.in
    #   uvicorn
uvloop==0.19.0
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.25.1
    # via pre-commit
watchfiles==0.21.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==12.0
    # via uvicorn
wheel==0.42.0
    # via pip-tools
wrapt==1.16.0
    # via
    #   testcontainers
    #   testcontainers-core
zstd==1.5.5.1
    # via asynch

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
